debos: Add support to create an image for development
debos is a tool to make the creation of various Debian-based OS images
simpler. It expects a YAML file as input and will run the actions listed
in the file sequentially, running the debos/debimage.yaml introduced it
will create a image ready to flash on a Beaglebone Black suitable for
oz2 development.

Signed-off-by: Enric Balletbo i Serra <eballetbo@gmail.com>