ozdisplay / ozmaps: Introduce a new function to find resources path
The purpose of find_datapath function is try to find where are located
the resources (images and maps). For now, it assumes that the executable
is in the 'bin' directory and the resources are in 'share/oz2'
directory. The datapath variable points to the base path of bin and
share directories.E.g:

 /usr
  |---local
      |---bin
      |---share
          |---oz2
              |---images
              |---maps

Signed-off-by: Enric Balletbo i Serra <eballetbo@gmail.com>