ozmaps: Remove DEBUG define from initmaps.c
Initially many oz programs were created with a main() routine built
within the program for unit testing. The #define DEBUG was used to
compile the main() routine (or not compile it if commented out).

The #define DEBUG concept along with main() routines for unit testing
were removed from the oz system in an earlier commit "76e0f7e Remove all
main functions under DEBUG define". This is a cleanup from that commit.

Signed-off-by: Peter Macleod Thompson <peter.macleod.thompson@gmail.com>