clang-format: Add configuration file
clang-format is a tool to format C/C++/...  code according to a set of
rules and heuristics.  Like most tools, it is not perfect nor covers
every single case, but it is good enough to be helpful.

In particular, it is useful for quickly re-formatting blocks of code
automatically, for reviewing full files in order to spot coding style
mistakes, typos and possible improvements.  It is also handy for
sorting ``#includes``, for aligning variables and macros, for reflowing
text and other similar tasks.  It also serves as a teaching tool/guide for
newcomers.

Signed-off-by: Enric Balletbo i Serra <eballetbo@gmail.com>