move find_datadir() to SDL2init.c
This is a preference issue.   As the "game" increases in functionality
and ddisplay from dieface daemons is integrated, displaymap.c will
increase in complexity.  The idea is restrict displaymap.c
to "game activity" only.  Hence the proposed move of find_datadir().

We also might want to move ALL external variables to maps.h or some
other include file - globalvar.h.   Also, an error reporting macro
that reports filename and linenumber for each error printed is a
possible addition.  I saw one on
https://stackoverflow.com/questions/3485150/file-and-line-number-for-debugging
comments?....

Signed-off-by: Peter Macleod Thompson <peter.macleod.thompson@gmail.com>